project_chatbot/requirements.txt
streamlit==1.28.0
pandas==2.1.0
requests==2.32.0
python-dotenv==1.0.0
fastapi==0.111.1
pydantic==2.6.0

# LangChain
langchain==0.1.0  # or latest compatible version

# Groq LLM
langchain-groq==0.0.6  # adjust if your version differs

# HuggingFace embeddings
sentence-transformers==2.2.2
transformers==4.35.0
torch>=2.1.0

# Optional: for vectorstores
langchain-core==0.0.175
